// frontend-angular/src/app/pages/profile/types/profile.types.ts
export interface MyProfileDto {
    name: string;
    lastName: string;
    phoneNumber: string;
    country: string;
    city: string;
    state: string;
    zipCode: string;
    title: string;
    summary: string;
    available: boolean;
    availableText: string;
    education: string;
    educationStartYear: string;
    educationEndYear: string;
    photoUrl: string;
    videoUrl: string;
    email: string;
    linkedInUrl: string;
    gitHubUrl: string;
}